+++ Assets/Scripts/Weapons/MachineGun.cs
@@ -2,9 +2,9 @@
 
 public class MachineGun : Weapon
 {
-    public int energyCost = 0;
+    public float energyCost = 0;
 
-    public override int Click(Transform transform)
+    public override float Click(Transform transform)
     {
         if (CanFire())
         {
@@ -25,7 +25,7 @@ public class MachineGun : Weapon
         return 0;
     }
 
-    public override int GetEnergyRequirement()
+    public override float GetEnergyRequirement()
     {
         return energyCost;
     }
